requirements.txt
python-dotenv>=1.0.0
gigachat==0.1.42.post1
pyTelegramBotAPI>=4.25.0
yandex-cloud-ml-sdk>=0.1.0
